[FEAT] : ⚙️ dialog fragment 분리
