+++ app/src/main/java/com/example/alldayfit/dietrecord/DietRecordFragment.kt
@@ -9,6 +9,7 @@ import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import android.provider.MediaStore
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -34,8 +35,6 @@ import java.io.File
 
 class DietRecordFragment : Fragment() {
     private var _binding: DietRecordFragmentBinding? = null
-    private val PICK_IMAGE_REQUEST = 1
-    private var dialogBinding: DietRecordAddDialogBinding? = null
     private val binding get() = _binding!!
     private val dietRecordsList: MutableList<String> = mutableListOf()
     private lateinit var dietRecordChart: BarChart
@@ -51,7 +50,6 @@ class DietRecordFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         _binding = DietRecordFragmentBinding.inflate(inflater, container, false)
-        dialogBinding = DietRecordAddDialogBinding.inflate(inflater, container, false)
         initView()
 
         dietRecordChart = binding.dietRecordChart
@@ -105,10 +103,6 @@ class DietRecordFragment : Fragment() {
 
     private fun initView() = with(binding) {
         /* 식사 위젯, 칼로리 수치 아이템 text 초기 설정*/
-        breakfastView.mealTxt.text = getString(R.string.diet_record_breakfast)
-        lunchView.mealTxt.text = getString(R.string.diet_record_lunch)
-        dinnerView.mealTxt.text = getString(R.string.diet_record_dinner)
-        snackView.mealTxt.text = getString(R.string.diet_record_snack)
 //        carbohydratesView.analysisTxt.text = getString(R.string.carbohydrates)
 //        proteinView.analysisTxt.text = getString(R.string.protein)
 //        fatView.analysisTxt.text = getString(R.string.fat)
@@ -133,7 +127,6 @@ class DietRecordFragment : Fragment() {
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
-        dialogBinding = null
     }
 
     /* main_graph의 action을 활용해서 dialog 띄우기 */
@@ -143,152 +136,5 @@ class DietRecordFragment : Fragment() {
 
     companion object {
         fun newInstance() = DietRecordFragment()
-    }
-
-    // 식단 기록의 ImageButton을 클릭했을 때 식단 기록 추가 다이얼로그를 띄우는 함수
-    private fun showDietRecordAddDialog() {
-        dialogBinding?.let { dialogBinding ->
-
-            val dialogView = dialogBinding.root
-
-            val builder = AlertDialog.Builder(requireContext())
-                .setView(dialogView)
-
-            val alertDialog = builder.create()
-
-            // RecyclerView 및 어댑터 초기화
-            val recyclerView = dialogBinding.mealListview
-            val dietRecordAdapter = DietRecordAdapter(dietRecordsList)
-
-            recyclerView.layoutManager = LinearLayoutManager(requireContext())
-            recyclerView.adapter = dietRecordAdapter
-
-          // 식단 입력 후 plus 버튼 눌렀을 때 음식이 추가되는 클릭 리스너
-          dialogBinding.btnAdd.setOnClickListener {
-              val dietRecord = dialogBinding.mealEdit.text.toString()
-              //editText가 빈칸일 때 toast 메세지 띄우고 식단 추가 되지 않음
-              if (dietRecord.isNotEmpty()) {
-                  dietRecordsList.add(dietRecord)
-                  dietRecordAdapter.notifyDataSetChanged()
-                  Toast.makeText(requireContext(), "식단 추가 완료!", Toast.LENGTH_SHORT).show()
-                  //식단 기록 입력 후 editText 초기화
-                  dialogBinding.mealEdit.text.clear()
-              } else {
-                  Toast.makeText(requireContext(), "식단을 입력해주세요.", Toast.LENGTH_SHORT).show()
-              }
-          }
-
-            // 식단 입력 후 plus 버튼 눌렀을 때 음식이 추가되는 클릭 리스너
-            dialogBinding.btnAdd.setOnClickListener {
-                val dietRecord = dialogBinding.mealEdit.text.toString()
-                //editText가 빈칸일 때 toast 메세지 띄우고 식단 추가 되지 않음
-                if (dietRecord.isNotEmpty()) {
-                    dietRecordsList.add(dietRecord)
-                    Log.d("yjRyu", "dietRecordsList = $dietRecordsList")
-                    dietRecordAdapter.notifyDataSetChanged()
-                    Toast.makeText(requireContext(), "식단 추가 완료!", Toast.LENGTH_SHORT).show()
-                    //식단 기록 입력 후 editText 초기화
-                    dialogBinding.mealEdit.text.clear()
-                } else {
-                    Toast.makeText(requireContext(), "식단을 입력해주세요.", Toast.LENGTH_SHORT).show()
-                }
-            }
-
-            dialogBinding.closeBtn.setOnClickListener {
-                alertDialog.dismiss()
-            }
-
-          dialogBinding.finishBtn.setOnClickListener {
-              if(dietRecordsList.isEmpty()){
-                  Toast.makeText(requireContext(), "식단을 입력해주세요.", Toast.LENGTH_SHORT).show()
-              } else {
-                  alertDialog.dismiss()
-              }
-          }
-
-          //권한체크 후 갤러리 열어주는 코드
-          dialogBinding.dietImg.setOnClickListener{
-              if (ActivityCompat.checkSelfPermission(
-                      requireContext(),
-                      Manifest.permission.READ_MEDIA_IMAGES,
-                  ) != PackageManager.PERMISSION_GRANTED
-              ) {
-                  (ActivityCompat.requestPermissions(
-                      requireActivity(),
-                      arrayOf(
-                          Manifest.permission.READ_MEDIA_IMAGES,
-                      ),
-                      1
-                  )
-                          )
-              } else {
-                  openGallery()
-              }
-          }
-          alertDialog.show()
-
-      }
-            dialogBinding.finishBtn.setOnClickListener {
-                alertDialog.dismiss()
-            }
-
-            dialogBinding.dietImg.setOnClickListener {
-                //Todo 권한체크 코드
-                openGallery()
-            }
-            alertDialog.show()
-        }
-
-    }
-    private fun openGallery() {
-        try {
-            val intent = Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI)
-            intent.type = "image/*"
-            startActivityForResult(intent, PICK_IMAGE_REQUEST)
-        } catch (e: Exception) {
-            // 갤러리 열기 중에 문제 발생 시 처리
-            e.printStackTrace()
-            Toast.makeText(requireContext(), "갤러리를 열 수 없습니다.", Toast.LENGTH_SHORT).show()
-        }
-    }
-
-    // 상대경로를 절대경로로 바꿔주는 코드
-    fun getRealPathFromURI(uri: Uri): String {
-        val buildName = Build.MANUFACTURER
-        if (buildName.equals("Xiami")) {
-            return uri.path!!
-        }
-        var columnIndex = 0
-        val proj = arrayOf(MediaStore.Images.Media.DATA)
-        val cursor =
-            (activity)?.contentResolver?.query(
-                uri,
-                proj,
-                null,
-                null,
-                null
-            )
-        if (cursor!!.moveToFirst()) {
-            columnIndex = cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DATA)
-        }
-        val result = cursor.getString(columnIndex)
-        cursor.close()
-        return result
-    }
-
-    // 이미지 선택 결과 처리
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == PICK_IMAGE_REQUEST && resultCode == RESULT_OK && data != null && data.data != null) {
-            val imageUri: Uri = data.data!!
-            val newImageUri = getRealPathFromURI(imageUri)
-
-            // Glide
-            Glide.with(requireContext())
-                .load(File(newImageUri).path)
-                .centerCrop()
-                .into(dialogBinding!!.dietImg)
-
-        }
     }
 }
